Add home view functionality, small changes in the database and daos

+++ src/main/java/sk/upjs/ics/daos/sql/SQLUserDao.java
@@ -181,7 +181,7 @@ public class SQLUserDao implements UserDao {
     public void updateBalance(User user) {
         String updateQuery = "UPDATE users SET credit_balance = ? WHERE id = ?";
         try (PreparedStatement pstmt = connection.prepareStatement(updateQuery)) {
-            pstmt.setLong(1, user.getCreditBalance());
+            pstmt.setFloat(1, user.getCreditBalance());
             pstmt.setLong(2, user.getId());
             pstmt.executeUpdate();
         } catch (SQLException e) {
